Move when we charge for packets to after validation (#196)
* Don't charge users for late packets

especially when they have multi-buy set to 1

* break out multi-buy adjustment

If we've seen a packet before, check if we want to buy more before
declaring it a duplicate

* Remove already handled case branch

This noop will get caught in validate_frame

* only set multi_buy the first time we validate a packet

* Validate seen frames

If a frame has made it this far and we've seen it before, multi_buy was
already set to accept more

* cleanup

* Handle replay / late packet correctly

- picked from macpie

* Bad merge, remove double join_timeout

* Initialize fcnt to -1

First frame comes through as 0 which causes us to think it's a
late_packet right out of the gate.

* Update test_utils for new device_worker state

* Fix formatting

* Test util helpers to get channels worker and event manager in tests

* Testing not buying late packet outside window

Originally:
A packet would come in just after we had handled it, and was no longer
in the frame cache, so it would be bought again.
Still being inside the RX window, 2 DC and 2 hotspots would show for a
packet that should have required 1 DC and 1 hotspot.

Now:
A device worker now tracks the FCnt, to recognize late packets better.

Well:
If a double packet comes in during the window, that would be
`router_device_routing`'s job to kick it out. We don't go through
`router_device_routing` in tests; if a packet slips in under the window
and we've seen it before, the device will buy it. That is okay.

* Move just a tad more outside the window

Previous runs, there was still a chance that a packet could sneek in
under the window if the RNG-CPU overlords were feeling it.

I'm still getting intermittent failures, but I can't get them to trigger
when logging is on, so this is a fun one for tomorrow.

* Wait until we're sure the next packet will be late

The timer and sending messages is who knows how accurate during testing
on different boxes. We know what makes a late packet though, once the
fcnt has been updated.

* Downcase logging for consistency

Co-authored-by: macpie <pierre@helium.com>
Co-authored-by: macpie <macpie@users.noreply.github.com>