Add counter to device updates
Some orgs have lots of devices, and when the logs are streaming by it's
hard to tell how many things are pending.

This could also help us determine if orgs are triggering updates while
updates are already going.

There's around ~9k devices on prod (as of this writing). We update 1
device at a time to keep from hammering console. That could take around
2.5 hours to update everything