Update cflist for AS923 regions
Originally took the frequencies from
https://www.thethingsnetwork.org/docs/lorawan/frequency-plans/#as920-923-as1

https://github.com/helium/miner/pull/772/files#
updated the frequencies list for the miners based off
https://lora-alliance.org/wp-content/uploads/2021/05/RP-2-1.0.3.pdf

They look to be based off the frequency offset for each sub-region.
But the document says frequency offsets don't apply to MAC commands and
cflists sent from servers. So we're going to send the same list that
maps to the miners lists when offsets are applied