Always send base64 payloads for queue updates (#320)
We're attempting to simplify the interactions with queuing downlinks
through console. Queuing through an endpoint or payload_raw we assume
the user is giving us a base64 encoded payload, router decodes and
stores, then sends that to Console (when queue updates are active).

Queueing downlinks through Console are base64 as well, sometimes.
To simplify we're building the contract that router will send whatever
it received back to Console, no need to think about who encoded or
decoded when or where.