Michaeljeffrey/ch9996/joinaccept dlsettings with as923 updates (#331)
* Update location to use new agreed asia suffix

This file might be removed soon, but until it is, let's keep it up to
date. eventually all this information will be encoded on the chain.

* Update default AS region until we can remove

* Placeholders for cflist of further AS regions

* Wrap Region in public functions to map to non sub-region versions

* merge join2 and rx2 windows info rx2_rf function

Move some record stuff around so the regions stick off the end a bit
more, makes this file a bit easier to navigate

* Update cflist for AS923 regions

Originally took the frequencies from
https://www.thethingsnetwork.org/docs/lorawan/frequency-plans/#as920-923-as1

https://github.com/helium/miner/pull/772/files#
updated the frequencies list for the miners based off
https://lora-alliance.org/wp-content/uploads/2021/05/RP-2-1.0.3.pdf

They look to be based off the frequency offset for each sub-region.
But the document says frequency offsets don't apply to MAC commands and
cflists sent from servers. So we're going to send the same list that
maps to the miners lists when offsets are applied

* TxQ records can't be constructed without .time field

Dialyzer knew something was wrong, but couldn't tell me quite what. rx1
windows hide this in the `tx_offset/4` function.

* One more update to cflists

relevant part of the spec:
AS923_FREQ_OFFSET only applies to end-device default settings. AS923_FREQ_OFFSET
does not apply any frequencies delivered to end-device from network server through MAC
commands or the CFList.

Whatever frequencies gets sent to the device, it will respect. So we
should send preshifted frequencies.

* Fix datarate used when queueing downlinks (#318)

We track the uplink datarates for devices. We were not using the mapped
datarate for downlinks when seeing if a payload can be queued. Uplinks
are tighter on size than downlinks are in most cases.

* Always send base64 payloads for queue updates (#320)

We're attempting to simplify the interactions with queuing downlinks
through console. Queuing through an endpoint or payload_raw we assume
the user is giving us a base64 encoded payload, router decodes and
stores, then sends that to Console (when queue updates are active).

Queueing downlinks through Console are base64 as well, sometimes.
To simplify we're building the contract that router will send whatever
it received back to Console, no need to think about who encoded or
decoded when or where.

* Add new data to cayenne decoder (#319)

* Add new data to cayenne decoder

* Remove useless comment

* Update netid parsing (#321)

This will help with metrics.

* Upgrade core (#323)

* Macpie/2021.05.27.0 fix (#324)

* Add s3_base_url

* Update deps

* Fix comma (#325)

* Disable ASan in CI (#326)

ASan causes rustc to segfault. It was a well known problem
when JEMalloc was pervasive in rust. However, I don't know
if rustc itself still links against jemalloc.

* Macpie/should update filters (#328)

* Fix should_update_filters when routing info not there yet

* Upgrade core

* Macpie/fix cts (#329)

* Bump core

* Bump blessed_snapshot_block_height

* Fix CT monthly_reward

* Expose Region default rx2 datarate index

Fixes #327

Ref: DLSettings from LLS 1.0.4.pdf
https://lora-alliance.org/resource_hub/lorawan-104-specification-package/

* DLSettings is now a binary, it knows how many bits it is

* Test each AS923 subregion individually

add timeouts to keep tests from never exiting

Devices don't know about subregions

start executable for top level region

export top_level_region

* f2uch does care about subregions

Map AS923 frequencies to what was sent down in the CFList.

* Remove lora_region function

Proto was updated, we didn't go with the `_ASX` format, and plain
`AS923` is not allowed.

* DataRate Index for TLR

Co-authored-by: macpie <macpie@users.noreply.github.com>
Co-authored-by: Jay Kickliter <jay@helium.com>