Michaeljeffrey/ch10024/dedupe xor filter (#344)
* dedupe new filters

* Add test for dupes added later

I think this case was already handled. We check if eui pairs are
_already_ in a filter. In the case of an update, that will be true.

If someone tries to add 2 devices with the same pair in between filter
updates is when the problem occurs.

* Also dedupe updates within filter

It's not just new filters, it's anytime duplicates are being added to
the same filter.

I added a version of uniq that preserves sort order. I don't believe
we get enough adds to impact performance, but there's room for
improvement.

* replace uniq with lists:usort/1

expand out some of the eunit asserstions to hopefully make it clear that
the Add argument is supposed to be in order and uniq.