Unique eui pairs coming from console db
If there were duplicate devices coming from Console and the filter they
already existed in was chosen for the new updates to go into.
(that filter was the smallest)

The xor hashing would break.

The duplicate eui pairs were being correctly marked as "already in a
filter", but _both_ pairs were marked that way. So when the updates came
through and duplicates were found in the smallest filter, attempting to
construct the new filter would have duplicates in it again.

Solution:
Dedupe the eui pairs we get from console as early as possible.

Duplicates eui pairs are allowed, there's still the app_key to
differentiate devices, and Console has a constraint on the triple
{dev_eui, app_eui, app_key} being unique. We just need to know where to
route joins.