Fix sending updates to console for matching devices
If a user adds a device that is already in a filter, we should not
commit an update to that filter, we should let Console know the device
has been added.

We only want to update filters when the underlying EUIs have changed,
but we still want to let console know the status of all devices by their
ID.

Good find Victor.