Michaeljeffrey/reflow xor filters (#378)
* commit so we don't lose anything

the fear is real

* test for removing devices from xor filter

* Here's the test that tests the problem

How do we handle unknown devices after a restart?

* add option to fail early out of waiting

* working tests

* remove txn_count for testing

* sort binaries by size

* more cleanup and specs

* more cleanup

* cleanup logging, more helpers

* A note so we don't forget

* remove redundant tests

* We're storing all filters, even if they're empty now

* [fixup] remove test

* add reporting functions

* Specs

* Update tests

- reverse order so if we pull from the head of a list we always get a
device from that round of updates
- remove anon functions
- more printing with reporting functions
- test reporting function

* first pass at cli file

* remove logging, move to explicit function calls

* move rebalance to cast, it's part of the api

* Don't crash the whole server if we can't make an API call

* spec distribute function

* appease the dialyzer

* match spec to order of docs

* refresh cache helper

* Sync device state rather than just write

If we wait until all txns are cleared, then we miss possibly removing
devices from previous txns. There is room for optimization in writing to
disk here.

* Remove extra logging for tests

Didn't really help anyways :sigh:

* incorrect spec for current state

* don't know when this changed, but unnecessary

* Track device_id through xor filter

This will let us send more information about xor filters to places like
Console.

* Diff to be added

Less writing to disk. Can double duty to sent do Console

* Send updates to console about added-removed devices

* We don't want to think we're unsetting

* split out diffing state with persisting it to places

* get_device_updates helper

Let's us build up future updates without committing them into filters
and losing information about what's happening

* Use new return from estimate_cost for reporting

* add test for estimate_cost

* Add cli_xor_filter to clique registry

- Add some more commands for initializing

* fix typo

That'll do it. Can't copy something that doesn't exist

* Should have caught this

I was just here

* remove xor filter update from device cli

* fix cli for updates to xor filter worker

* Make sure lists with low numbers print

* Split commiting into 2 steps

- add rebalance
- rename init -> reset_db

* Fix formating

* udpate cache when resetting db

* Actually empty rocksdb

* make extra sure to assign filter_index when balancing

* Make sure maps are default same size when diffing

* Add newline to update printout

* Get eunit tests working again

Lots of type changes

* Fix dialyzer and eunit

* Some fixes

* Update filter_report_device

* Fix db in filter_report_device

* Fix get_fold

* Don't break esimate_cost when there's no updates

* updating to cache and db work with lists now

The devices contain the filter they belong to, we don't need to pass
redundant information around in the maps that result in more changes of
the data's shape for no reason.

* Fix sending updates to console for matching devices

If a user adds a device that is already in a filter, we should not
commit an update to that filter, we should let Console know the device
has been added.

We only want to update filters when the underlying EUIs have changed,
but we still want to let console know the status of all devices by their
ID.

Good find Victor.

* Use actual console URL for xor filter updates

* Don't crash when asking about timer that's running

While running, the reference for the timer has destroyed itself, and
returns false rather than 0 time.

* update to match device cli interface

I think this is easier

* There seems to be some cpu contraints on github

if things are slow, the chain can be not updated when we ask for the filters

* Move all sleeps to expect_block

Tests keep breaking on lines after expect_block in github. low cpu power :(

* Reduce printout for blockchain_worker

not usefule in these tests anyways

* Fix multi txn test

We used to expect a txn for filter 0 because things were different. But
not that we're not submitting txns that would be duplicates of what
exists, we only go forward 4 blocks.

* fix spec

Co-authored-by: macpie <pierre@helium.com>