Michaeljeffrey/migrate xor filter (#389)
* Don't cause filter txns for remove only filters

Removing devices from filters is not as important as adding. We can
remove old devices from a filter when its been chosen to be the filter
new devices are added to.

* Migrate filter

adds function for taking all devices in a filter and merging them with
devices in another filter, emptying the source filter. This operation is
manual, and will allow us to learn how filters fill up in different environments.

* make sure filter indexes are numbers coming through the cli

* Add dry run heading correctly

* Up nonce by count

Groups may not have all filters present, so nonce would skip by filter choice

* make sure all devices know what filter they're going into

* Allow all filters to be considered for smallest

not just filters we know have devices

* Make BaseEmpty with 0-index aware seq

we consider only as candidates the number of fitlers we're allowed to
have, says the chain.