Add buffer calc to sc expiration
Router attempts to always have at least 1 safety valve state channel
open. That paired with the expiration calculation lets us get in a
position where we have more than one state channel that closes within a
few blocks of each other.

It appears this is causing latency spikes when mulitple channels are
closed close together. This is an attempt to have some minimum amount of
time between closes to prevent bursts of packets that have been waiting
to get into a state channel.

note: exhausted state channels can not close, they have to wait for
their expiry time.

It might be a good idea to brainstorm a different way to think about
state channel expirations, router is no longer in the position where its
bouncing back and forth between two channels, and the purpose of the
second was the be an immediate overflow when the first needs to close.

There are cases where many more state channels are opened because they
max out actors. Depending on who makes it into which state channels, the
rate at which each channel drains is not something we currently track or
compensate for.