Potential crash when sending events to console
When asking for a channel we assume all frequencies in events going to
console are uplink frequencies. Most of the  time this is okay because
uplinks and downlinks follow more or less the same line. That's not the
case for all regions.

If we can't get a channel assuming an uplink frequency, try assuming
it's a downlink frequency, otherwise we have a good failure on our
hands.

fixes #379