Remove device from db and cache if posting an event returns 404
If a device becomes unlinked from from an org somehow and router does
not catch that, it will continue to pull it from the database when
offers come in with the matching EUI or devaddr. Nothing can really get
rid of the device now because the supervisor didn't _actually_ start
anything, so there's nothing to kill, and we don't have any db sweeps to
clean out old devices.