microsoft azure iot hub pre built integration (#434)
* Azure integration

* Working get/create/delete

* Move azure http to azure_connection module

* Remove actually printing, can bring back if there's problems

* consolidate and simplify azure connection

* HTTP and MQTT use different URI for tokens

* flesh out mqtt api

* flesh out azure channel to behave like aws channel

* move to generating tokens and passwords when used

Don't want to keep around an old token then start getting rejected
because it's expired

* handle downlink from azure

* Add ping to azure mqtt

* moar binary

* switch format order

Easier to find the url at the front, we don't know how big the body is
going to be, so it's hard to find the url in the logs.

* log when we get a channel we don't recognize

* convert azure channel from api

* add message identifier and reason to test_utils

* return with status to we know to schedule ping

* default headers not exported

* add azure a channel option in the tests

* pull out connection information

this allows us to mock things easier while testing :dancer:

* Azure channel test suite

* update credential names to keep in line with aws

* Update from master merge

- we added hold_time to some of the console messages

* use correct api version for mqtt

- export password generators for repl testing

* Resource uri needs to be encoded

* need to use encoded version for both parts of the sas

* use pythons default ping time from msft docs

* up expiration time, missing publishes

* add more logging to failed mqtt

* formatting

* go back to shorter ping

understand more of the keepalive-ping relationship, that was not the issue

* use unique_id for channel

the lesson we learned from the bare mqtt integration

* don't use eunit api tests until we have a mock in place

These tests were used while exploring the azure http endpoints. They're
good to have around. We should get a mock in place like we have for console.

* log warning for unrecognized responses

good looking out prof

* make sure devices are iotEdge enabled when creating

I thought this would fix the mqtt connection issue where Azure is
closing the connection regardless of pings and pongs. It did not,
probably still good to have.

* rig up to connect and publish, then disconnect

azure connection is not reliable yet, so we're going to connect and
publish then disconnect. unfortunately this means we don't have a
reliable way to handle downlinks right now.

* y.k. we need regular devices for sending messages

* Make sure trailing slash is in uplink topic

- Go back to regular connection strat

* bring back the downlink part of tests

the connections have been stable for a bit now, let's keep these parts
of the azure suite alive!

Co-authored-by: macpie <pierre@helium.com>