Consider channel_correction in calculating frame timeout
When channel correction is to be part of a downlink, our frame timeout
needs to be the short window so we have a chance of getting the downlink
to the device within the RX window.

It would be nicer to ask the frame if there are any FOpts going down,
but we don't know that when we're calculating the timeout. It could be
moved to before hand, but that will take some different work.