Add another xor filter migrate cli command
move to front.
Event balancing the xor filters was a bad idea because you get charged
by the size of the filter, not the size of the diff you're causing.
If you get a few thousand devices in a filter, you can start making
updates that cost 3mil DC. That's not very sustainable.
So the idea is you move most of your devices to the front filters, then
slowly add more devices to the end/cheaper filters. When those start to
get large and expensive, you compact them into the front ones again.