XOR filter test investigation (#526)
* device_to_json only for devices from ets

the default device in console callback is already in the desired JSON format.

* fix typo

* After exhausting console devices, put them back

Some tests (xor in particular) rely on being to treat Console endpoints
like a stable store. When we're paginating the results put all the
devices back so we can do it again.

There could still be race conditions if the tests are ever parallelized.