Remove a channels backoff when it's deleted (#547)
* add types for channel and backoff maps

* add helper for getting integer from application env

* Remove backoffs for old channels

If a channel failed to start it does not make it's way into the
channel_map. It does however have an entry in the backoff_map.

When the channels are updated, we need to make sure we're not holding on
to any old backoffs who's messages could come through and attempt to
restart a channel that was removed.