Add support for indexing into lists in mustache template
When we come across a list where the first item is a map, we assume the
rest of the list is full of maps and we create a 0-indexed map of that
list.

This allows the renderer to continue looking for a value.
You cannot terminate a template value _on_ a list, and you cannot index
into lists of lists.