Regex is the answer to the 2-door riddle
bbmustache gives us some fun flexibility by allowing for functions to be
passed in lieu of data. We use this to add some functions to allow
casting into different types. However, once you return from that
function with _not_ a function, bbmustache will carry on its merry way
forgetting you were ever even around.

This wouldn't be so bad, if bbmustache was not a sneaky sneaky
sneakster. bbmustache does not let you know when it's dealing with
"sections" (conditionals or loops). That's the secret sauce, how it
makes itself useful, if it didn't you might end up reimplementing the
whole thing, and then where would bbmustache be?

Sections are what breaks being able to treat every list of maps
like an indexed map of maps. The `{{#data}}{{/data}}` style things stop
working. Pay no mind to the library behind the curtain!

Mixed metaphors aside,,,
Here to save the day like a loan-shark comes Regex.

We can use it to drop little breadcrumbs of what's to come. bbmustache
will be handing us our own clues without ever knowing. We looked ahead for
keys that are accessing lists with indexes and marked them. Then we can
be sure to treat the list of maps like a map of maps with confidence.

Which brings us to the answer of the 2-door riddle. The only winning
move is to not play bbmustache's game. It's a strange game. Maybe some
chess instead.