Add support for indexing into lists in mustache template (#575)
* Add support for indexing into lists in mustache template

When we come across a list where the first item is a map, we assume the
rest of the list is full of maps and we create a 0-indexed map of that
list.

This allows the renderer to continue looking for a value.
You cannot terminate a template value _on_ a list, and you cannot index
into lists of lists.

* Regex is the answer to the 2-door riddle

bbmustache gives us some fun flexibility by allowing for functions to be
passed in lieu of data. We use this to add some functions to allow
casting into different types. However, once you return from that
function with _not_ a function, bbmustache will carry on its merry way
forgetting you were ever even around.

This wouldn't be so bad, if bbmustache was not a sneaky sneaky
sneakster. bbmustache does not let you know when it's dealing with
"sections" (conditionals or loops). That's the secret sauce, how it
makes itself useful, if it didn't you might end up reimplementing the
whole thing, and then where would bbmustache be?

Sections are what breaks being able to treat every list of maps
like an indexed map of maps. The `{{#data}}{{/data}}` style things stop
working. Pay no mind to the library behind the curtain!

Mixed metaphors aside,,,
Here to save the day like a loan-shark comes Regex.

We can use it to drop little breadcrumbs of what's to come. bbmustache
will be handing us our own clues without ever knowing. We looked ahead for
keys that are accessing lists with indexes and marked them. Then we can
be sure to treat the list of maps like a map of maps with confidence.

Which brings us to the answer of the 2-door riddle. The only winning
move is to not play bbmustache's game. It's a strange game. Maybe some
chess instead.

* fix dialyzer

for our purposes with kvc, proplists and maps work the same.

* We don't have `Data` we have a function wrapping the Data

* Print the NormalMap if there's an error

Data was a function, and printing that doesn't provide much help when
coming across this error message. From the NormalMap, we can recreate
what happened.

* Make sure we're not trying to index into a string

This adds support for arbitrary lists of data

* Throw on context misses

Otherwise, if you try to access a key that doesn't exist, mustache will
simply insert nothing.

* More tests and explanations

- test for out of bounds indexing
- test for really large lists
- test for index simple list of data