merge device and api metadata before rx_delay bootstrap
the api endpoint that gives us the device for joins has a paired down
version of the metadata (don't want to give up too much info for
potentially not the right device(?)) so we need to merge it with the
existing device metadata to make sure we don't lose any information.

The expected flow is,,,
device is updated with an rx_delay, router gets a device refresh
message, that happens pretty quickly, then a new join comes through and
we pick up from the previous metadata updates.