Preferred Hotspots (#456) (#730)
* Add preferred_hotspots to router_device

This commit adds the preferred_hotspots field to router device,
bumping the version of the record to v7.

* Update device routing to support preferred hotspots

Device selection logic remains the same.  After a device is selected:

  - if the device has no preferred hotspots, the packet is sent as before

  - if the device has preferred hotspots, the the packet will be sent
    only if it was received from a preferred hotspot.  Multibuy logic
    is bypassed in this case.

* Use correct routing info format in test join packets

* Add preferred hotspots to router_console_api

Also add to console_callback for testing.

* Add preferred hotspot checks for packets

This commit adds preferred hotspot checks under handle_packet/3.  It
includes some dialyzer cleanup as well.

* Remove redundant EUnit tests

This commit removes a number of eunit tests related to preferred
hotspots.  Since this functionality depends on so many side effects,
it is perhaps better tested in CT.

* Add CT tests for preferred hotspots

* Move preferred hotspots to metadata

This commit removes the preferred_hotspot field from the device record
and places it in the device's metadata instead.  The device record is
bumped from v7 back to v6.