Revert OTP version for docker until TLS is fixed (#739)
* Revert OTP version for docker until TLS is fixed

- [TLS handshake failure after upgrade to 24.3.4](https://github.com/erlang/otp/issues/5961)

* Upgrade and Pin grpcbox & chatterbox until stable from blockchain-core

blockchain-core is still in the middle of updating grpc communication.
grpcbox and chatterbox were moved from andymck org to novalabsxyz,
orphaning the ref in the lock file. core is not quite ready for another
bump, but we need to be able to pull these two repos to continue building.

* upgrade blockchain, now includes up to date grpc deps

* more bumping. grpc tests no longer fail

I think I got the udpate between merges. either way, grpc deps and tests
need to be considered very soon.