update erlang-lorawan library (#773)
The initial erlang-lorawan library caused issues with Downlinks.

One core issue: there was a typo in the list of DataRates. Somehow we missed adding the SF7BW500 datarate. That list determines the up to down datarate on the Downlink RX1 Window.

The added erlang-lorawan unit tests catch the issue. The tests call both the legacy API and the new library API and do a direct output comparison.