Use `undefined` as the default fcnt for devices
This commit updates router_device so that the fcnt field is either
undefined or non_neg_integer().  The value is undefined until the
device successfully joins and sends the first frame.  After the first
frame is successfully processed, the value is 0.  Thereafter, the
value is incremented after each successfully processed frame.