Update device worker to use FCnt from routing
This commit updates router_device_worker to accept a frame count
argument in calls to accept_uplink/6 and handle_frame/9.  The worker
no longer considers the FCnt field of a frame packet but instead
relies on this cryptographically verified argument from routing.

On a successful join, the worker now sets the device's fcnt field to
undefined instead of -1.