[Issue #643] Implment the uplink frame counter as a 32-bit integer (#780)
* Add router_device:fcnt_next_val/1

This commit adds fcnt_next_val/1 to router_device and updates
router_device_routing to use it.  Router will now maintain a 32-bit FCnt.

* Use strings instead of binaries for default testing datarate.

* Use `undefined` as the default fcnt for devices

This commit updates router_device so that the fcnt field is either
undefined or non_neg_integer().  The value is undefined until the
device successfully joins and sends the first frame.  After the first
frame is successfully processed, the value is 0.  Thereafter, the
value is incremented after each successfully processed frame.

* Update routing to use the an expected 32-bit FCnt

This commit updates routing in the following ways:

1.  The device record's fcnt field is now maintained as a 32-bit
counter.
2.  Only packets with the correct "current" or "next" fcnt values will
be routed.  In other words, only the next packet or retransmits of the
previous packet will be allowed through.
3.  After a successful join, routing will expect the first frame to
have fcnt=0 but will tolerate fcnt=1.
4.  If a packet is successfully routed, the FCnt value of that packet
is sent to the device worker.

* Update device worker to use FCnt from routing

This commit updates router_device_worker to accept a frame count
argument in calls to accept_uplink/6 and handle_frame/9.  The worker
no longer considers the FCnt field of a frame packet but instead
relies on this cryptographically verified argument from routing.

On a successful join, the worker now sets the device's fcnt field to
undefined instead of -1.

* Use frame counters correctly in router discovery

* Use frame counts starting at 0 in router_SUITE

* Update router_device_worker_SUITE for new FCnt argument

 - Add the FCnt argument to calls to accept_uplink/6.
 - Count frames from 0.

* Simplify return values.  Part of dialyzer cleanup.

* Expand 32-bit rollover test a bit.

* Fix bad formatting in a log statement.

* Save the updated FCnt in device_worker instead of device_routing.

This commit moves the logic for saving the device record with updated
FCnt to the device worker.

* Ensure fcntdown is limited to 32 bits

This commit adds router_device:fcntdown_next_val/1 which ensures
fcntdown values are limited to 32-bit values; router_device_worker is
also updated to use this function in all places where fcntdown is incremented.

* Remove strict ordering of frame counters from routing

This commit changes the logic of device routing to remove (as much as
possible) the requirement that end device's send strictly ordered
frame counters.

* Test cleanup and formatting

* Reduce the number of channel refreshes when saving updated FCnt

* Fix a race condition in router_xor_filter_SUITE

* remove orphaned code

Co-authored-by: Michael Jeffrey <michaeldjeffrey@gmail.com>