Fix encoding of preferred hotspots in Console events (#811)
* Format preferred hotspots in Console events

* Use console_callback to set preferred hotspots in tests

This commit updates some of the integration tests for preferred
hotspots to use console_callback to set preferred hotspot values
instead of pre-loading the device cache.  This is a more correct
approach.

Also, a bug in `handle_packet_from_strange_hotspot` is corrected: the
test should have exercised the case where the end device joined via a
preferred hotspot and then router receives a packet from a
non-preferred hotspot.