Always mount MAC commands into events, not so with integrations.
Any time MAC commands are present in an event, we should send them to Console.

In the case of downlink_acks, the downlink wasn't generated from an
integration, so we don't put fake integration information into the event
in hopes of not causing confusion on the User side they see an
integration name they didn't create.