Don't replace index lookup when no brackets are present
In cases where strings have periods followed by numbers, but none of the
mustache interpolation parts (some encoded string with periods and
numbers) we don't want to think it's an indexed lookup unless there are
brackets present.

We check to make sure the brackets are closed.

If someone using an encoding that allows curly brackets, and somehow
they properly open and close each other, all bets are off, that UB.