add a configurable delay while sending large batches of updates (#949)
* add a configurable delay while sending large batches of updates

If too many items are sent to the config service, it will get stuck
reading everything from the network before it will continue doing any db
operations, this can be very slow (think minutes). If we added a few
strategic pauses to our streaming up of items, the config service can
keep up and hopefully not fall behind, making the request seem to take
only as long as it takes to send everything up.

* fix typo

* add eunit

start fold with 1 so we don't incur a sleep on the first iteration.