ets table per devaddr
inserting devices into a ets bad where the whole object needs to be
unique slows down severely as the table grows but the keys do not grow
with the table. Inserting 250k devices spread across 32 devaddrs was
approaching 5 minutes.

making a table per devaddr introduces a bit of indirection, but ets sets
are fast, and device ids are unique.