manage fcnt (#1003)
* don't set max to 0 for these cases

That would refuse all future copies of this packet because they exceeded
the multi-buy. We want the rest of the already purchased packets to come
through okay.

* multi buy is primarily done in hpr

don't re-check here, the packet was bought, let's use it

* don't blanket update device fcnt from the packet

This already happens deeper in this call graph when the packet is
accepted and used

* remove 32 bit rollover test

the max 32bit frame count is somewhere in the 4billions.
the highest frame count for a device in use here is ~1,800,000.
136 years has about 4billion seconds.

* return verified_fcnt when finding device for data

The device_worker will be provided with the verified fcnt rather than
the expected fcnt when handling a packet.

* differentiate finding device for join and data

* remove inactive devices that are chosen during routing

* export as helper

* rename, uplink is more easily recognizable over data

* bring back updating the device fcnt

A verified frame count is being provided here rather than a speculative
frame count.